Stabilize map filter bar browser tests
